# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stripe Service API",
    "description": "Payment processing service built on Stripe. Handles checkout sessions, payment intents, webhook processing for payment events, and integrates with runs-service for cost tracking.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://stripe.mcpfactory.org",
      "description": "Production"
    },
    {
      "url": "http://localhost:3011",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Service-to-service API key"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "nullable": true,
            "description": "Additional error details"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "CreateProductResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "productId": {
            "type": "string",
            "description": "Stripe Product ID"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Product description"
          }
        },
        "required": [
          "success",
          "productId",
          "name"
        ]
      },
      "CreateProductRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Custom Stripe Product ID for idempotent creates"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "Product name"
          },
          "description": {
            "type": "string",
            "description": "Product description"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Custom metadata"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreatePriceResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "priceId": {
            "type": "string",
            "description": "Stripe Price ID"
          },
          "productId": {
            "type": "string",
            "description": "Associated Stripe Product ID"
          },
          "unitAmountInCents": {
            "type": "number",
            "description": "Price amount in cents"
          },
          "currency": {
            "type": "string",
            "description": "Currency code"
          }
        },
        "required": [
          "success",
          "priceId",
          "productId",
          "unitAmountInCents",
          "currency"
        ]
      },
      "CreatePriceRequest": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "minLength": 1,
            "description": "Stripe Product ID to attach the price to"
          },
          "unitAmountInCents": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "description": "Price amount in cents"
          },
          "currency": {
            "type": "string",
            "default": "usd",
            "description": "Currency code (default: usd)"
          },
          "recurring": {
            "type": "object",
            "properties": {
              "interval": {
                "type": "string",
                "enum": [
                  "day",
                  "week",
                  "month",
                  "year"
                ],
                "description": "Billing interval"
              },
              "intervalCount": {
                "type": "integer",
                "minimum": 0,
                "exclusiveMinimum": true,
                "default": 1,
                "description": "Number of intervals between billings"
              }
            },
            "required": [
              "interval"
            ],
            "description": "Recurring pricing configuration. Omit for one-time prices."
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Custom metadata"
          }
        },
        "required": [
          "productId",
          "unitAmountInCents"
        ]
      },
      "CreateCouponResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "couponId": {
            "type": "string",
            "description": "Stripe Coupon ID"
          },
          "name": {
            "type": "string",
            "nullable": true,
            "description": "Coupon display name"
          },
          "percentOff": {
            "type": "number",
            "nullable": true,
            "description": "Percentage discount"
          },
          "amountOffInCents": {
            "type": "number",
            "nullable": true,
            "description": "Fixed amount discount in cents"
          },
          "currency": {
            "type": "string",
            "nullable": true,
            "description": "Currency code"
          },
          "duration": {
            "type": "string",
            "description": "Coupon duration"
          }
        },
        "required": [
          "success",
          "couponId",
          "name",
          "percentOff",
          "amountOffInCents",
          "currency",
          "duration"
        ]
      },
      "CreateCouponRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Custom Stripe Coupon ID for idempotent creates"
          },
          "name": {
            "type": "string",
            "description": "Display name for the coupon"
          },
          "percentOff": {
            "type": "number",
            "minimum": 1,
            "maximum": 100,
            "description": "Percentage discount (1-100). Provide this OR amountOffInCents."
          },
          "amountOffInCents": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "description": "Fixed amount discount in cents. Provide this OR percentOff."
          },
          "currency": {
            "type": "string",
            "description": "Currency code. Required if amountOffInCents is provided."
          },
          "duration": {
            "type": "string",
            "enum": [
              "once",
              "repeating",
              "forever"
            ],
            "default": "once",
            "description": "How long the coupon applies (default: once)"
          },
          "durationInMonths": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "description": "Number of months (required if duration is repeating)"
          },
          "maxRedemptions": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "description": "Maximum number of times the coupon can be redeemed"
          },
          "redeemBy": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 date after which the coupon can no longer be redeemed"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Custom metadata"
          }
        }
      },
      "GetProductResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "productId": {
            "type": "string",
            "description": "Stripe Product ID"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Product description"
          }
        },
        "required": [
          "success",
          "productId",
          "name",
          "description"
        ]
      },
      "GetPriceResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "priceId": {
            "type": "string",
            "description": "Stripe Price ID"
          },
          "productId": {
            "type": "string",
            "description": "Associated Stripe Product ID"
          },
          "unitAmountInCents": {
            "type": "number",
            "nullable": true,
            "description": "Price amount in cents"
          },
          "currency": {
            "type": "string",
            "description": "Currency code"
          },
          "active": {
            "type": "boolean",
            "description": "Whether the price is active"
          }
        },
        "required": [
          "success",
          "priceId",
          "productId",
          "unitAmountInCents",
          "currency",
          "active"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/products/create": {
      "post": {
        "summary": "Create a Stripe product",
        "description": "Creates a product in Stripe. Pass an optional id for idempotent creates — if the product already exists, it will be returned.",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID (injected by workflow-service)"
            },
            "required": false,
            "description": "Campaign ID (injected by workflow-service)",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID (injected by workflow-service)"
            },
            "required": false,
            "description": "Brand ID (injected by workflow-service)",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow name (injected by workflow-service)"
            },
            "required": false,
            "description": "Workflow name (injected by workflow-service)",
            "name": "x-workflow-name",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/prices/create": {
      "post": {
        "summary": "Create a Stripe price for a product",
        "description": "Creates a price attached to a Stripe product. Use the returned priceId in checkout sessions.",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID (injected by workflow-service)"
            },
            "required": false,
            "description": "Campaign ID (injected by workflow-service)",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID (injected by workflow-service)"
            },
            "required": false,
            "description": "Brand ID (injected by workflow-service)",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow name (injected by workflow-service)"
            },
            "required": false,
            "description": "Workflow name (injected by workflow-service)",
            "name": "x-workflow-name",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePriceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Price created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePriceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/coupons/create": {
      "post": {
        "summary": "Create a Stripe coupon",
        "description": "Creates a coupon in Stripe. Pass an optional id for idempotent creates — if the coupon already exists, it will be returned.",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID (injected by workflow-service)"
            },
            "required": false,
            "description": "Campaign ID (injected by workflow-service)",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID (injected by workflow-service)"
            },
            "required": false,
            "description": "Brand ID (injected by workflow-service)",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow name (injected by workflow-service)"
            },
            "required": false,
            "description": "Workflow name (injected by workflow-service)",
            "name": "x-workflow-name",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCouponRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Coupon created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCouponResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/products/{productId}": {
      "get": {
        "summary": "Get a Stripe product by ID",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "productId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID (injected by workflow-service)"
            },
            "required": false,
            "description": "Campaign ID (injected by workflow-service)",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID (injected by workflow-service)"
            },
            "required": false,
            "description": "Brand ID (injected by workflow-service)",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow name (injected by workflow-service)"
            },
            "required": false,
            "description": "Workflow name (injected by workflow-service)",
            "name": "x-workflow-name",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Product found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductResponse"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/prices/{priceId}": {
      "get": {
        "summary": "Get a Stripe price by ID",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "priceId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID (injected by workflow-service)"
            },
            "required": false,
            "description": "Campaign ID (injected by workflow-service)",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID (injected by workflow-service)"
            },
            "required": false,
            "description": "Brand ID (injected by workflow-service)",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow name (injected by workflow-service)"
            },
            "required": false,
            "description": "Workflow name (injected by workflow-service)",
            "name": "x-workflow-name",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Price found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPriceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Price not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/prices/by-product/{productId}": {
      "get": {
        "summary": "List active prices for a product",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "productId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID (injected by workflow-service)"
            },
            "required": false,
            "description": "Campaign ID (injected by workflow-service)",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID (injected by workflow-service)"
            },
            "required": false,
            "description": "Brand ID (injected by workflow-service)",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow name (injected by workflow-service)"
            },
            "required": false,
            "description": "Workflow name (injected by workflow-service)",
            "name": "x-workflow-name",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Prices for product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPricesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/coupons/{couponId}": {
      "get": {
        "summary": "Get a Stripe coupon by ID",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "couponId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID (injected by workflow-service)"
            },
            "required": false,
            "description": "Campaign ID (injected by workflow-service)",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID (injected by workflow-service)"
            },
            "required": false,
            "description": "Brand ID (injected by workflow-service)",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow name (injected by workflow-service)"
            },
            "required": false,
            "description": "Workflow name (injected by workflow-service)",
            "name": "x-workflow-name",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Coupon found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCouponResponse"
                }
              }
            }
          },
          "404": {
            "description": "Coupon not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}